Core/dhcp pxe boot (#35)
* Corrected tiny typo

* Pass host IP, gateway, subnet and netmask to docker containers as system variables

* Initial WiP changes for PXE Boot installation method

* Initial WiP Frontend changes for PXE Boot installation method

* typo fixes to swagger file.

* rename ansible examples

* Updated Ansible examples to include PXEBoot

* Cosmetic changes

* Added tftpd-hpa

* Fixed path to dhcpd run file

* Added /tftpboot volume

* Simplified logic in line with simplified dhcp_template - only hostname has to be uniqe

* Simplified template - subnet declaration not needed for each host entry

* Mount KSDIR as volume

* Removed unnecesary symlink

* tftpd-hpa config

* TFTPBOOT paths cleanup

* Generate dhcpd.conf on startup if applicable and prepare tftpboot structure on ISO upload

* Prepare tftpboot structure on ISO upload and PXE/tftpboot job cleanup

* tftpboot directory

* Not using kea dhcp so removing this file

* Adding syslinux 3.86 files to repo so no need for symlinks

* Adding syslinux 3.86 files to the repo for simplicity and portability

* Removed old script for getting host VM IP address

* v2.0 with PXE boot installation method

* Added macaddr, netmask and gateway columns to EAISTATUS table

* Start tftpd from run_isc-dhcp-server.sh script

* Update Initial Setup sections with Python dependencies

* Added pre-requisites check

* Added loggin for API jobs PUT endpoint and status codes 10-19

* Added Cancelled state

* Changed message on PXE boot ready state

* Error on POST method when job with same hostanme and Ready to deploy status already exists

* Corrected 2.0 release date

* Updated JobStatus schema and status dictionary

* Logging and some code cleanup

* Updated for PXE boot installation

* Code cleanup

* Fixed mainlog logger handle call

* Updated Readme for DHCP and fixed typos.

* Added FAQ about SecureBoot

* Corrected installmethod

* Corrected installmethod check

* Removed FAQ CIMC Install does not install ESXi

* Fix for IP validation with port number

* Fix GUI can use multiple MAC address formats

* Fix install failed because http state update error

+++ auto-installer_docker/run_docker.sh
@@ -1,9 +1,51 @@
-echo killing old docker processes
+echo Checking pre-requisites
+# check pre-requisites: docker-compose
+which docker-compose >/dev/null
+if [ $? -ne 0 ]
+then
+    echo "[ERROR] Missing docker-compose binary - install the package and re-run this script. Aborting."
+    exit 1
+fi
+# check pre-requisites: python3
+which python3 >/dev/null
+if [ $? -ne 0 ]
+then
+    echo "[ERROR] Missing Python binary - install the package and re-run this script. Aborting."
+    exit 2
+fi
+# check pre-requisites: python3
+pip3 show netifaces >/dev/null
+if [ $? -ne 0 ]
+then
+    echo "[ERROR] Missing Python netifaces library - install and re-run this script. Aborting."
+    exit 3
+fi
+# check pre-requisites: python3
+pip3 show jinja2 >/dev/null
+if [ $? -ne 0 ]
+then
+    echo "[ERROR] Missing Python jinja2 library - install and re-run this script. Aborting."
+    exit 4
+fi
+
+# (re)building docker containers
+echo Killing old docker processes
 docker-compose rm -fs
 
-echo export host IP address
-export EAI_HOST_IP=`python3 /opt/eai/get_host_ip.py`
-echo $EAI_HOST_IP
+echo Exporting host network settings
 
-echo building docker containers
+NETWORK_DATA=( `python3 get_host_network_settings.py` )
+export EAI_HOST_IP=${NETWORK_DATA[0]}
+export EAI_HOST_GW=${NETWORK_DATA[1]}
+export EAI_HOST_SUBNET=${NETWORK_DATA[2]}
+export EAI_HOST_NETMASK=${NETWORK_DATA[3]}
+
+echo
+echo EAI_HOST_IP: $EAI_HOST_IP
+echo EAI_HOST_GW: $EAI_HOST_GW
+echo EAI_HOST_SUBNET: $EAI_HOST_SUBNET
+echo EAI_HOST_NETMASK: $EAI_HOST_NETMASK
+echo
+
+echo Building docker containers
 docker-compose up --build -d
